python-django-uwsgi/requirements.txt
Django==2.2.2
djangorestframework==3.9.4
mysqlclient==1.4.2.post1
pytz==2019.1
sqlparse==0.3.0
uWSGI==2.0.18
